Optimise image_render_interpolate_icc.
Split out the core of the routine into a function call. The
guts of the original routine go into irii_inner_template,
which can be static instantiated as required.

I had hoped that merely instantiating this 3 times would be
enough to optimise the routines, but in the end I had to
manually cut the core down for the 2 cases that the AirPrint
test file hits.