Make pdfimage devices work on Linux
Because the devices usurp the FILE * created by the gdevprn code and
turn it into a stream, we must set the stored pointer to NULL after
closing or freeing the stream in order to avoid gdev_prn_close()
attempting to close the underlying file again, which causes an error
on Linux.

Also remove some routines we no longer use.