Define gs_restrict as a synonym for the compiler 'restrict' definition
Using the 'restrict' tag tells the compiler that it can safely assume
no aliasing of pointers is involved, allowing for even higher levels
of optimisation. However since this is a non-standard extension
(though widely supported) the exact definition varies.

We know of 'restrict', '__restrict', '__restrict__' and 'cdecl(restrict)'

The Visual Studio cdecl causes problems with us simply using 'restrict'
and #defining it in stdpre.h because stdlib.h includes the cdecl. If
we include stdpre.h before stdlib.h then the compiler tries to expand
'cdecl(resgister)' to 'cdecl(__register)' and that fails.

The only way to accommodate all the various options is to use something
other than the 'restrict' keyword, and to #define that as required for
the different compilers.

This commit introduces the 'gs_restrict' keyword and that is #defined
in stdpre.h to be appropriate for whatever compiler is in use (the
default is to define it to nothing). This means that wherever we want
to use the restrict capability, we must include stdpre.h.