pdfimage - change the way files are closed
The 'all devices' regression test persists in throwing errors and
seg faulting with these devices when closing. I can't reproduce the
problem on Linux.

So instead of preventing the gdevprn code from closing the file, prevent
the stream code from closing the file instead, and allow the gdevprn
code to perform normally.

Hopefully this will solve the problem.....