Fix pattern code for devices handling patterns by subclassing new device
The pattern code permits for us to use an 'internal' accumulator or for
a device to declare it can handle patterns itself. When a device handles
patterns we pass spec_ops to the device to delineate the pattern
operations.

It can happen that when the pattern terminates we (g)restore back to a
point before the internal pattern accumulator device was pushed into
the graphics state. To handle that possibility we push a pointer to the
device onto the exec stack and use that stored pointer to tell the
pattern accumulator device when the pattern terminates, even if it is
no longer the current device at that time.

Unfortunately this conflicts with the ability for other devices to
handle patterns. The code uses that same stored device pointer to
indicate whether we are using the internal accumulator or not. If the
stored pointer is NULL then we are not using the internal accumulator.

But this assumed that, if we aren't using the internal accumulator, that
the current device is the one which declared it could handle the pattern.
This is dangerous anyway, because there's not restriction on how fat
back we might have restored, and its possible (if unlikely) that the
device now in the graphics state isn't the one that can handle
patterns. In addition, if the underlying device handles patterns by
subclassing itself with a new pattern handling device, then when we
restore back, even though the device is the one which stated it could
handle patterns, its not the one we want to inform of the pattern
termination.

We handle this simply by getting the current device after the 'start'
notification. I think its reasonable to assume that the device now in
place is the one which expects to receive the 'close' message. We
then store that device on the exec stack. When we get to the cleanup
routine, if we decide we are not using the internal accumulator, then
we call the new stored device to handle it.

I think ideally we should store a flag for whether we are using the
internal accumulator, but it wouldn't be any more efficient.