pdfimage - support the '%d' output format specifier
It hadn't occurred to me that anyone would want to do this, so I hadn't
implemented it. On reflection its probably useful, and it should
(finally) solve the all devices regression report problem.

At the same time, add the file to the Windows solution so that we get
debug information in it.