pdfwrite - drop unusable Matte from SMask when doing colour conversion
Bug #698817 "ColorConversionStrategy=Gray breaks image with smask"

SMask images can have a Matte entry which is a preblended colour. It
must have the same number of components as the parent image. If we are
doing colour conversion then we can't unconditionally use the existing
Matte value, as it may have the wrong number of components for the
converted image.

In fact Matte values are optional and rare. Even rarer are Matte entires
which have any actual effect. We can't use the value (unless the parent
image was already in the target colour space, and we can't realistically
convert the Matte to another space.

So here we test the color conversion strategy, if the image colour space
is not the same as the target, then we simply discard the Matte entry.
This isn't 'correct', but then nothing will be, and this stops Acrobat
from whinging and refusing to render the image at all.