XPS interpreter - when discarding an ICC profile for an image, NULL hte pointer
Possibly Bug #698834 " xps dump"

Its not at all clear what the bug reporter sees as a problem, but when
I run the file to the display the XPS interpreter warns that it is
discarding an ICC profile because it has the wrong number of
components (profile has 4 in, 3 out components, image claims to be
RGB).

This counts the profile reference down by 1 but doesn't set the pointer
to the profile, saved in the image structure, to NULL leading to us
trying to free a garbage pointer later.

Setting the pointer to NULL solves the problem, and may fix the bug
report, if that's what the reporter is complaining about.