Fix problem with bitrgbtags and ROPs.
With bitrgbtags, we can get called to do 24 bit rop operations
with transparency, with colors like 0x01ffffff.

The transparency rop code is supposed to spot 0xffffff as meaning
"be transparent", but the tag in the top byte confuses it.

Accordingly, mask off stray bits in the colors before using them.

The same potentially happens for an 8 bit + tags device so we cope
with that too.