Fix stale pointer de-reference during GC due to restore.
If a CIEBasedDEF or DEFG colorspace was set in a save level, a restore
would free the Table memory, but the pgs->icc_profile_cache still had
a (stale) pointer to that area. Fix by copying the Table to memory that
is not subject to save/restore (stable_memory). Rely on the GC to free
up the previous table data since we have no idea what allocator was
used for it (from PostScript it would be local or global VM). Seen with:
-Z:?@$ -r300 -sDEVICE=psdcmyk -o x.psd tests_private/comparefiles/Bug688308.ps