ROP fix: Correctly set scolors/tcolors S and T data case.
We weren't setting S and T colors in this particular case of
the code, which would have shown errors if we'd ever called
it with S and T being 1 bit data arrays to choose between
scolors[0] and [1] etc.

It appears that no code in our entire regression suite actually
calls the S and D both being data case anyway, as proved by
cluster testing this code with an "abort();" line inserted!