PS interpreter - don't use findresource for Idiom Recognition in bind
The current implementation of idiom recognition uses findresource in
the definition of the bind operator to check all existing IdiomSets
for a candidate to substitute.

This is, in fact, wrong. The PLRM (p135 3rd Edition) states that :

"This matching by value occurs only for IdiomSet instances that are
defined in VM; bind does not consider instances that are in not in VM
but only in external storage."

and:

To ensure that the instances in VM are consistent with the external
ones, the interpreter automatically invokes findresource to load
external IdiomSet instances into VM at the beginning of each job and at
certain other times."

We were not loading the IdiomSet resources into VM, and we were searching
external resources during bind.

Not only is this incorrect, but because findresource ends up calling
filenameforall on every directory in the search path, its also very
slow.

While normally we don't expect to see many occurences of bind during the
course of a job, the customer supplied file did multiple binds on every
page, and had > 50,000 pages. This led to a fourfold decrease in
performance compared to Acrobat Distiller.

Here we load the external resources into gobal VM at startup, and we
now *only* search instances in VM. This is more correct going by the
spec, and improves the performance by the desired factor of 4 for the
customer. It provides a small benefit on cluster testing, which is
expected as this only affects PostScript input.

No differences expected.