Fix devn color affecting psdcmyk tests/pdf/Bug6901014_launch_leaflet.pdf p2
The image interpolation uses copy_alpha_hl_color so when InterpolateControl=-1
is used with this file in clist mode, playback ends up calling the
default_no_copy_alpha_hl_color function which throws an "undefined" error. If
this proc is seen, instead pass to copy_alpha which can work as long as the
devn color can be packed into a chunky color which is the case as long as the
num_components is <= 8, which is useful enough to make it worth trying. If
the num_components is too large, the copy_alpha will throw a "rangecheck".