pdf_info.ps - gracefully handle circular Resource references
Although not strictly part of Ghostscript, pdf_info.ps is used by
customers and this is reasonably easy to improve.

When checking for transparency, pdf_info.ps uses the Ghostscript
PDF interpreter to do the work. However, other aspects of its operation
(Media Sizes, Font use) aren't available from the PDF interpreter.

When detecting Font usage, pdf_info.ps walks the Page Resource chain,
and any the Resource chains of any Pattern or XObject Resources used
on that page, as well as the Resources for any Annots (and also any
XObject or Pattern Resources used by the Annots). If any of these
Resource chains contains a circular reference (which they absolutely
should not) then pdf_info will enter a loop and process unwil it
exceeds the execution stack limit, throwing an error.

In this commit we borrow the technique used by the PDF interpreter, we
create a dictionary and every object which we inspect has its object
number stored in that dictionary. Whenever we encounter a new object
we first check to see if its number is in that dictionary. If it is, we
know we've seen it before and this is (possibly) a circular reference.

Even if it is not a circular reference, we've already checked that chain
so we don't need to check it again.

This isn't cluster checked.