Fix opening of files > 2GB on Windows built with VS after VS2013
For whatever reason, even on 64 bit builds, Visual Studio changed
behaviour and the compiler/run-time now invoke __fstat64i32 for
fstat. This returns an error if the file size can't fit in int32
(the older fstat just returned a negative number, ignoring overflow)
Change to use __fstat64 which works on VS 2005 so it should be
fine. I also tested this build on an ancient Dell with 760Mb RAM
running Windows XP SP3 and it runs fine, so __fstat64 is not a
problem going back that far.