pdfwrite - Fix the detection of type 3 descendants in composite fonts
Bug #698844 "ps2pdf creates unusually large PDF files starting with version 9.21"
Bug #698647 "PCL5 - Downloadfont not as T3 Font interpreted"
Bug #607462 "Ghostscript falls into infinite loop with pdfwrite"

The actual problem was that a Type 0 composite font with a single
descendant was used with stringwidth and the space glyph. This was not
detected by pdf_text_begin() and so the glyph was cached, but it was not
added to any type 3 font, it was simply processed as a stringwidth.

Later the same glyph from the same font was used in earnest, pdfwrite
noticed it didn't have a copy of the glyph and set up to capture the
CharProc. But the glyph was present in the glyph cache which meant the
CharProc wasn't executed, leading to an infinite loop.

The fix applied prevented caching type 3 fonts, but that caused problems
with cases where we wanted to convert a font to type 3, so the fix
was extended to only prevent caching of input type 3 fonts, not
synthesised ones.

However, this still caused problems with programs which used type 3 fonts
with glyphshow. Because we did not cache the glyphs we would detect each
glyph use through glyphshow as a new glyph, and so there would be no
glyph reuse, leading to large slow PDF files.

Here we extend the check on descendants of type 3 fonts so that if no
font change takes place, we still check the 'current' descendant to
see if its type 3. If a type 3 descendant is detected then we skip
caching the glyph which prevents the whole problem.

No differences expected.