Fix bug 698823. Segfault with psdrgb and transparency_example.ps
The psdrgb device has max_components == 3, but the pdf14 device was
using spot_colors to set it more than that (transparency_example.ps
doed not set PageSpotColors so it was -1, leading to underflow).
If page_spot_colors is the default -1, then don't use it, and also
limit the num_components to max_components for the device.