Reduce severity of memory debug problem detection
Bug #698849 "gs_abort with Bad object"

The error occurs because of the combination of -Z? and -dLastPage. In
order to do First/Last Page processing, we insert a subclassing device
ahead of the regular device (ppmraw in this case).

The subclassing works by making a new device structure, copying the
existing device to it, and replacing the content of the existing
device structure. This is done to maintain coherency of gsave/grestore
and save/restore; the structures maintain pointers to the device in
the graphics state structure, so we can't simply move it.

When subclassing, we change the declared size of the structure (ssize)
but we don't (can't!) change the size that the memory manager allocated
in the first place. This leads to a mismatch between the declared size
of the structure and the allocated size. I mentioned this in the
original commit for subclassing, but noted that in normal use there is
nowhere that actually checks.

However, if you set -Z? then additional checks are done on memory
validity, and this discrepancy is flagged. It also raises an abort,
presumably because this is seen as serious enough to stop and
investigate.

I don't want to remove this check altogether, as it may still be useful,
but we don't want to abort now because this might be a benign change
caused by subclassing.

So here, pull the size != ssize check out separately from the other
consistency checks, and if it is not true raise a warning, but do
*not* abort the process.