Bug 697856: Reduce clist buffer memory usage for worst case files.
The band_height is chosen based on number of bytes per scan line, and
a fudge factor if the page uses transparency. For pages with a large
number of raster lines, the number of bands could cause the space
needed for the band "states" to exceed the space remaining in the
buffer. This was handled by doubling the BufferSpace until there
was enough space for the band states. Improve by increasing the space
available more gradually. For one of the worst case files, Bug689581.pdf,
this reduces the memory needed from 40Mb to 24Mb.