Bug 694260: FAPI: tidy up error return for broken TT glyph
In the event of an error retrieving a TTF glyph, we were storing the negative
error code in the same variable we use to store the glyph length for a valid
glyph, which is an unsigned variable, causing a valgrind error, and potential
security issue reading uninitialized memory.

In fact, other consumers silently cope with this type of problem and
(presumably) use a notdef glyph instead. So this tweaks the error codes to
do the same.

Because we now have to differentiate between different types of error, we
can no longer just use a negative value, so there is an enumerated type for
the error values.