Bug 698867: Don't push texttransparency knockout group for invisible text
The customer's example file had text that was not opaque, but was being
rendered in TextRenderingMode 3 (invisible). Don't push the extra
transparency group for this. Reduces time from 483 seconds to 44 seconds
with default clist params, and with BufferSpace of > 32m down to 14 seconds.
Also reduces page mode (-dBufferSpace=2000m) from 40 minutes to 9 seconds.