Fix Bug 698874 - incorrect text positioning.
The bug title is misleading, this a font selection problem where 1 of
2 fonts with the same attributes must be selected.  Experiments
suggest the HP chooses the font which was defined earliest.  We
emulate the same effect in our interpreter by creating sequential PJL
font numbers for downloaded fonts and the selection algorithm
prefering lower numbers.

Progression with the test file bug690520.pcl.