pdfwrite - take advantage of GSUB table glyph substitution enhancement
Bug #691574 "vert punctuation is rotated when convert ps to 1.3 pdf"

Commit 7dd033589ce40649ab602fba4c0473aa40c48439 was an enhancement to
apply vertical glyph substitution from a TrueType GSUB table. This is
primarily used when a user specifies a TrueType font to be used as a
replacement for a missing CIDFont (especially in PDF files). Using a
horizontal font as a replacement for a vertical font can result in
punctuation marks at the wrong orientation.

This commit uses the machinery added in that previous commit to apply
the same fix for pdfwrite. Note that this still does not produce
'correct' output, because its still a substitute font being used. It
does produce punctuation marks at the correct orientation now though,
for the limited selection of test files we have. The other relevant
report is 698471, though that was reported against rendering, the same
effect occurs with pdfwrite.

The cluster shows one diff, a progression, which surprises me since I
cannot duplicate the reference output, but I'm not going to complain
since its a progression.