pdfwrite - ensure we have a substream Resource dictionary before using it
Bug #692182 "xps -> pdfwrite SEGVs with softmasked pattern"

After much messing about with this bug over the years, it appears
that most of the bugs have been fixed. However one still remained;
we could attempt to add a ProcSet Resource to a substream, even
though we had not yet allocated a Resource for that substream.

Although it shows up with this file, technically it could
probably happen with any kind of input.

Fixed here by allocating a substream Resource dictionary if we
don't already have one when we want to add the Procsets to it.

No differences expected