Documentation - correct erroneous information in psfiles.htm
Psfiles.htm incorrectly listed a number of files as being in 'lib'
when in fact they have been moved to 'Resource/Init. In addition
Several files were listed which no longer exist, and some files in
'examples' were not listed or described.

I suspect this listing is still not comprehensive, but at I believe
the information that is there is now at least correct.