Bug 698767: Improve handling of multiple master fonts
The FAPI API didn't really handle the possibility that multiple master fonts
could change (via setweightvector) without a call to definefont/setfont etc.

We now allow two ways for FAPI to handle this: if the scaler in force allows it,
we'll set the new weight vector, and carry on as before.

If the scaler does *not* allow us to change the weight vector "on the fly",
we'll destroy the scaler's font object and related data, and recreate it with
the new weight vector in the font dictionary.