Add a TIFF specific CFLAGS and avoid discarded const warning
libtiff has a large table initialization with string literals that gcc
automatically makes const, but the structure definition for the table does not
qualify those entries as const. Hence we get warnings.

This adds TIFF_CFLAGS and, with a gcc compatible compiler, sets it to:
-Wno-write-strings

which kills that warning.

Done this way because we want to retain that warning in the Ghostscript code.