Device subclassing - fix potential problems
Flagged by scan build, the code to work a chain of devices had a
potential problem in walking up and down the chain.

Also flagged by scan build, the create_compositor method checked the
target child device before trying to use it to set the saved target
color info, but didn't check it before trying to set the procs.