squash GCC warnings
remove an unused variable from a (large) macro in gxfilltr.h

remove some unused variables, and an entire unused function in
sjpx_openjpeg.c

remove an unused variable in gdevbit.c

remove an unused variable in gdevjpeg.c

bracket a variable declaration and usage with #ifdef DEBUG to prevent
the compiler warning that the variable is set but not read on non-debug
builds. gdevxps.c

preserve 'const' when passing a memory pointer to a procedure to
avoid a warning about discarding const. zcie.c