Fix jmp_buf related compiler warning (setjmp_.h)
The compiler (gcc/clang) considers the alignment of entries in a structure to
be a run-time "variable", thus gives a warning when that value is used to
size an array used in a globally available data type.

The solution (suggested by Robin) is to use a buffer twice the size of a jmpbuf,
thus guaranteeing we have enough space to offset into the buffer, and get the
required alignment.