Remove the Mitsubishi CP50 colour dot matrix printer device
Recent changes caused this device to enter clist mode more frequently
and our automated testing revealed that in clist mode this device
does not work. It calls a device method which should never be called.

Its possible to remove the call, but we have no way to determine whether
the resulting output actually works.

Given the age of this device (and Mitsubishi have reused the model
number for a different device, so it must be old) we are removing
support for this device and deleting it fro the build.

If someone is still using this device, we'll reinstate support if they
are prepared to test the outptu for us.