Tweak bitmap interpolater patch size.
When we interpolate an image, we restrict the interpolation to
the "patch" of the destination image that is actually visible.
Previously, the calculation of this patch has been inaccurate,
resulting in a larger area being decoded than was actually
needed.

While this extra work was largely mitigated by the "Active" area
detection code, it has meant that the contribution calculations
within the scaler were 'wrapping' the outside pixels in the
region to avoid accessing out of bounds.

This upsets the new library we are working on to do the scaling,
and it's basically inelegant. Instead we fix the destination
patch calculations so that the patch is correctly sized based
upon the available source data. This is safe to do because we
know the source data was already expanded.