Remove the 'device filter chain'
This was an early attempt to allow the possibility of 'chaining'
devices together. I believe it was never actually completed, although
the stack existed and was carefully preserved in the graphcis state,
the only code which accessed it was '.popdevicefilter'.

No C code or PostScript code ever actually pushed a device onto the
device filter stack, and nothing ever examined the stack to pass
operations through the 'chain'.

Since its been superseded by device subclassing, and didn't actually
seem to do anything anyway, remove it altogether. In addition to
removing clutter this also reduces the amount of 'magic' surrounding
reference counted objects in gstates.