Remove faulty code picked up by scanbuild
scanbuild correctly warned that a variable was uninitialised. In fact
it looks to me like the code that was flagged simply can't work.

Fortunately it isn't exercised by any files in our test suite, and has
never been reported as a bug, so its clearly an uncommon (perhaps even
impossible) case.

All the same, lets fall back to the default handling if we hit it,
because at least that works, which the old code definitely didn't.