Remove an unused function, clear a scanbuild warning
The function cos_write_stream_from_pipeline was never used, so remove
its definition.

At the same time try to squash a harmless scanbuild warning. We us a
'while(s....' loop to try and find a stream, scanbuild regards this as
implying that 's' might be NULL *initially*, which is not true. This
leads to a 'NULL pointer dereference' warning in the line after the
loop, because scanbuild thinks we can get to s->state with s being an
inital NULL.

Since we already check for s becoming NULL inside the loop, we don't
need to check it in while condition.