Change from CMM_THREAD_SAFE #define to a gscms_is_threadsafe function
In preparation for enabling sharing the profiles and link cache among
rendering threads, switch from a compile time flag to a function that
is returned by the CMS interface module linked in (e.g gsicc_lcms2.c)

This avoids more build machinery to propagate the flag, and ensures
that the flag won't be mis-matched to the CMS in use.

Also flag a few FIXME places identified that relate to multi-threaded
sharing of profiles and the profile link cache. Initially, even though
lcms2art is thread safe, return "false" until the multi-threaded
rendering works.