Handle a potential error condition in PXL
Bug #698995 "Fixing error returns in gxicolor.c can cause PXL to enter an infinite loop or seg fault"

This started off as a fix to a scanbuild warning and spiralled somewhat.

Fixing gs_image_class_4_color() so that it didn't ignore an error return
led to the possibility of the PXL interpreter seg faulting or ending up
in an infinte loop.

Initially this turned out to be because gs_initgraphics() was no longer
setting the colour space to the proper default, leading to us trying
to use the very pattern space which failed. That was fixed in the
previous commit.

After that it turned out that in an error condition we were freeing
an image enumerator twice. This didn't seem to cause the PostScript
interpreter a problem, though it seems likely it did really, it simply
didn't exhibit before the interpreter exited. The PXL interpreter
however threw an error in the memory manager rather rapidly.

Fixed that by not calling gx_default_end_image in gx_enum_begin()
when we hit this particular error condition (no handler for an image).

This removes one of the free instances, and seems more reasonable anyway
it seems wrong to end an image we haven't yet begun.

This allows us to now propagate the error up from px_remap_pattern
without causing a seg fault or infinite loop. (finally)