Fix 694689 - Seg faults found by fuzzing.
Allow clients of the decompression routines to pass in a parameter to
indicate no further input or output is expected.  The decompression
routines should produce an error if the parameter is set and more
processing is needed.  The pattern code uses this to indicate indicate
the final scanline has been processed, it is not needed by the image
code.