More in this commit than planned!
1) Adds gs_main_set_device() and gsapi_set_device() calls so the Ghostscript
device can be set via the C API. Not as trivial as for the other languages
since it requires the Postscript world to be properly synced with the
internals.

2) Adds a proc_get_device_memory() API into the language switching
interpreter implementation structure.

3) Tweak how Ghostscript is initialized from the language switching code,
so the garbage collecting allocator is available when we want to create a
new device.

4) Add the code in ps_impl_set_device() to call the new API calls added in 1).

5) Have a debug build throw an assert should more than one interpreter supply
an allocator from the proc_get_device_memory() call.