PCL interpreter - correctly identify high level devices
Pointed out by Chris, the code in here explicitly tested the name of the
device to decide if it was a high level device. We deprecated that
approach some time ago, replacing it with a spec_op call instead.

The comment said 'this needs a better solution' and this is it.

Should work correctly now with any device which identifies itself as a
high level device.