pdfwrite - remove dead code
Commit efb9bf3949617b4c733a6dcce04acd142c50115d changed the flow of
control to avoid going through code which could not work, since it left
a variable unassigned (and didn't do any part of the work that would
be required to give the variable a sensible value).

That left some code unreachable leading to a coverity warning (Coverity
ID 262411). I was unable to find any way to trigger the original
condition, and the code as it stands should at least work, even if it
might be sub-optimal in some cases. So remove the dead code to prevent
the warning, if this needs fixing in future there will need to be more
code written anyway.