pdf_info.ps - check for null DescendantFonts after oforce
pdf_info.ps was modified to check for circular references when resolving
indirect objects. Now if we find recursion we return a null.

However, we weren't checking that when resolving DescendantFonts which
could lead to a typecheck error if we had two different CIDFonts on
the same page, and the CIDFonts used the same DescendantFonts array
(which is silly, of course).