PDF interpreter - fix ToUnicode processing of bfrange with high byte set
Bug #699041 "Cyrilic text not respected"

The title is a misnomer, the problem is not Cyrillic fonts as such.

The example file has an unusually complex ToUnicode CMap, which includes
bfrange entries where the CID has a value > 0xFF. The code in pdf_font.ps
to create our internal GlyphNames2Unicode table was adding the 'offset'
(The high byte of the index) to the CID, which basically means we were
adding the high byte twice (the key already included the high byte).

Fix this by simply not adding the offset to the keys.

Its surprisingly hard to find examples of this, I can't immediately see
any in our test suite. I have gone through some of the files which
have been reported as problems with ToUnicode processing in the past
but none of them fall into this condition.

A couple of files show difference, but these are irrelevant, they are of
the 'wrong before, differently wrong now' category.